Enforce contract naming requirement
Enforces the naming requirement to not include a '_' and also
verifies that the contract version is a valid semver string.

Also updates contract examples and integration tests to reflect
the newly implemented requirements:

Versions must be a valid semver string. The old versions in the
contracts, '1.0', were invalid semver strings. Updates to valid
semver strings, so '1.0' -> '1.0.0'.
Also updates the 'intkey_multiply' contract to be referred to as
'intkey-multiply'.

Signed-off-by: Shannyn Telander <telander@bitwise.io>